reconcilers/scheduler: turn locationdomain into locationstatus controller
